Femora/mora/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
firebase-admin>=6.4.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-chroma>=0.1.0
chromadb>=0.4.22
google-generativeai>=0.8.5
faiss-cpu>=1.7.4
pypdf>=4.0.1
sentence-transformers>=2.5.1
requests>=2.31.0
